backend/requirements.txt
# Web framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
authlib==1.3.0
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# GCP SDK
google-cloud-compute==1.15.0
google-cloud-storage==2.14.0
google-cloud-secret-manager==2.18.1
google-auth==2.27.0

# Utilities
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0

# Development
black==24.1.1
isort==5.13.2
mypy==1.8.0
